refactor: 폴더정리 및 접근제어자 수정
