+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/Bank.swift
@@ -8,9 +8,14 @@
 import Foundation
 
 struct Bank {
-    var bankManager: BankManager
-    var queue: CustomerQueue
-    var randomNumber: Int = 0
+    private var bankManager: BankManager
+    private var queue: CustomerQueue
+    private var randomNumber: Int = 0
+    
+    init(employee bankManager: BankManager, customer queue: CustomerQueue) {
+        self.bankManager = bankManager
+        self.queue = queue
+    }
     
     mutating func start() {
         while true {
@@ -29,7 +34,7 @@ struct Bank {
         }
     }
     
-    func displayMenu() {
+    private func displayMenu() {
         print(
 """
 1 : 은행 개점
@@ -38,7 +43,7 @@ struct Bank {
 """, terminator: " ")
     }
     
-    mutating func selectMenu() -> String {
+    mutating private func selectMenu() -> String {
         guard let input = readLine() else {
             return ""
         }
@@ -46,19 +51,17 @@ struct Bank {
        return input
     }
     
-    mutating func updateCustomerQueue() {
+    mutating private func updateCustomerQueue() {
         randomNumber = Int.random(in: 10...30)
         let numberList = Array<Int>(1...randomNumber)
         
-        
-        for i in numberList {
-            let customer = Customer(number: i)
-            
+        numberList.forEach {
+            let customer = Customer(number: $0)
             queue.enqueue(data: customer)
         }
     }
     
-    mutating func handleCustomer() {
+    mutating private func handleCustomer() {
         for _ in 0..<queue.count {
             guard let customer = queue.dequeue() else {
                 return
@@ -68,8 +71,9 @@ struct Bank {
         }
     }
     
-    func displayEndMessage() {
-        let totalHandlingTime = String(format: "%.2f", bankManager.handlingTime * Double(randomNumber))
+    private func displayEndMessage() {
+        let totalHandlingTime = String(format: "%.2f",
+                                       bankManager.handlingTime * Double(randomNumber))
         
         print("업무가 마감되었습니다. 오늘 업무를 처리한 고객은 총 \(randomNumber)명이며, 총 업무시간은 \(totalHandlingTime)초입니다.")
     }
