+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/BankManager.swift
